requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
banks==2.4.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
et_xmlfile==2.0.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.4.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
llama-index==0.14.20
llama-index-core==0.14.20
llama-index-embeddings-openai==0.6.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.7.5
llama-index-readers-file==0.6.0
llama-index-vector-stores-postgres==0.8.1
llama-index-workflows==2.19.0
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
openai==2.32.0
openpyxl==3.1.5
packaging==26.1
pandas==2.3.3
pgvector==0.4.2
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.13.2
pydantic_core==2.46.2
PyMuPDF==1.27.2.2
pypdf==6.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram==0.19.0
python-telegram-bot==22.7
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
striprtf==0.0.26
telegram_text==0.2.0
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
